app/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
django-unfold @ https://github.com/williamcpierce/django-unfold/tarball/main
django-mptt==0.14.0
django-import-export==3.2.0
django-auditlog==2.3.0
gunicorn==21.0.1
djangorestframework==3.14.0